requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
faker==30.1.0
pandas==2.2.3
